+++ apps/api-media/src/schema/mux/video/video.ts
@@ -129,11 +129,12 @@ builder.mutationFields((t) => ({
       if (!currentRoles.includes('publisher')) {
         where.userId = user.id
       }
-      await prisma.muxVideo.findUniqueOrThrow({
+      const video = await prisma.muxVideo.findUniqueOrThrow({
         where
       })
 
-      await deleteVideo(id)
+      // only delete mux asset if original user
+      if (video.userId === user.id) await deleteVideo(id)
 
       await prisma.muxVideo.delete({ where: { id } })
